I'm ok with this approach after looking around on other play Components.
All the construction is lazy, so to really start it on boot we need to access it.

After looking into a bunch of Components and examples of Compile time injection
this seems to be the way others are doing it. For example the evolutions
component and documentation is one of those examples.

https://github.com/playframework/playframework/blob/019020803e8a8dad2a66071a6379091a22d23b93/documentation/manual/working/commonGuide/database/Evolutions.md

https://github.com/playframework/playframework/blob/2.4.x/framework/src/play-jdbc-evolutions/src/main/scala/play/api/db/evolutions/EvolutionsModule.scala